Shard Python Authentication tests
Separates into two files in a new package
at test/python/auth. Divided by default
and external authenticators.
